Interview Task: Split 'Addons' and 'Views' pages
Modified app as follows:

- Split `Addons` and `Views` into separate pages

- Added links to these pages into side navigation in `app.vue`

- Added `starting_page` setting in `config.json` and code in `route.js`
  to re-route '/' to selected start page.

  - If `starting_page` not included in `config.json` then defaults to
    `home.vue`.

This commit also includes an *alternative* way to implement the default
start page — a call to `$f7router.navigate()` can be included in the
`mounted()` method of `home.vue` to re-direct. Included this as an
example but my suggested way to do it is in `routes.js` as that's where
I'd naturally look for default routes...

A new config file needs to be added to support these changes with the
following structure:

  {
    "apiKey" : "<supersecretcode>",
    "apiEndpoint" : "https://api.example.com",
    "apiSandboxEndpoint" : "https://api.example.com",
    "starting_page" : "<start_page>"
  }

where `<start_page>` is a string of the desired start page. Not included
as `config*` currently included in .gitignore (for obvious reasons).

Signed-off-by: Rod Manning <rod.manning@taslytic.com>